Sin refactorizar

+++ src/main/java/com/icm/telemetria_peru_api/dto/VehiclePayloadMqttDTO.java
@@ -12,7 +12,7 @@ public class VehiclePayloadMqttDTO {
     private Long companyId;
     private String licensePlate;
     private String imei;
-    private Integer speed;
+    private Double speed;
     private String timestamp;
     private Double fuelInfo;
     private Integer alarmInfo;
